requirements.txt
Flask
google-api-python-client
google-auth
yt_dlp
youtube-transcript-api